gst/va: move AVC encoder test classes to lib
