+++ lib/gstreamer/va/encoder.py
@@ -11,8 +11,10 @@ from ....lib.gstreamer.encoderbase import BaseEncoderTest, Encoder as GstEncoder
 from ....lib.gstreamer.util import have_gst_element, get_elements
 from ....lib.gstreamer.va.util import mapprofile, map_best_hw_format, mapformat
 from ....lib.gstreamer.va.decoder import Decoder
+from ....lib import platform
 from ....lib.codecs import Codec
 from ....lib.common import get_media, mapRangeInt
+from ....lib.formats import PixelFormat
 
 class Encoder(GstEncoder):
   @property
@@ -128,3 +130,43 @@ class EncoderTest(BaseEncoderTest):
 
   def map_profile(self):
     return mapprofile(self.codec, self.profile)
+
+############################
+## AVC Encoders           ##
+############################
+
+@slash.requires(*have_gst_element("vah264dec"))
+class AVCEncoderBaseTest(EncoderTest):
+  def before(self):
+    super().before()
+    vars(self).update(
+      codec         = Codec.AVC,
+      gstdecoder    = "vah264dec",
+      gstmediatype  = "video/x-h264",
+      gstparser     = "h264parse",
+    )
+
+  def get_file_ext(self):
+    return "h264"
+
+@slash.requires(*have_gst_element("vah264enc"))
+@slash.requires(*platform.have_caps("encode", "avc"))
+class AVCEncoderTest(AVCEncoderBaseTest):
+  def before(self):
+    super().before()
+    vars(self).update(
+      caps      = platform.get_caps("encode", "avc"),
+      gstencoder= "vah264enc",
+      lowpower  = False,
+    )
+
+@slash.requires(*have_gst_element("vah264lpenc"))
+@slash.requires(*platform.have_caps("vdenc", "avc"))
+class AVCEncoderLPTest(AVCEncoderBaseTest):
+  def before(self):
+    super().before()
+    vars(self).update(
+      caps      = platform.get_caps("vdenc", "avc"),
+      gstencoder= "vah264lpenc",
+      lowpower  = True,
+    )
